Fix job_v3 handlebar helper
Refactors this to reuse the mapJobClassV4toV3 function from
the job controller
