+++ src/jobs/interceptors/update-job-v3-mapping.interceptor.ts
@@ -7,6 +7,7 @@ import {
 import { Observable } from "rxjs";
 import { UpdateJobDtoV3 } from "../dto/update-job.v3.dto";
 import { UpdateJobDto } from "../dto/update-job.dto";
+import { mapUpdateJobDtoV3ToV4 } from "../job-v3-mappings";
 
 /**
  * PATCH/api/v3/jobs requires an UpdateJobDtoV3 object as request body.
@@ -19,24 +20,7 @@ export class UpdateJobV3MappingInterceptor implements NestInterceptor {
     const request = context.switchToHttp().getRequest();
     const dtoV3 = request.body as UpdateJobDtoV3;
 
-    let newBody: UpdateJobDto = {
-      statusCode: dtoV3.jobStatusMessage,
-      statusMessage: dtoV3.jobStatusMessage,
-    };
-
-    let newjobResultObject = dtoV3.jobResultObject;
-    // if executionTime is provided, add it to jobResultObject, to maintain compatibility
-    // after the job update is completed, it will be then moved to jobParams
-    if (dtoV3.executionTime) {
-      newjobResultObject = {
-        ...newjobResultObject,
-        executionTime: dtoV3.executionTime,
-      };
-    }
-    newBody = {
-      ...newBody,
-      jobResultObject: newjobResultObject,
-    };
+    const newBody: UpdateJobDto = mapUpdateJobDtoV3ToV4(dtoV3);
     request.body = newBody;
     return next.handle();
   }
